+++ lib/dio_cache_interceptor.dart
@@ -7,3 +7,4 @@ export 'src/model/cache_response.dart';
 export 'src/store/cache_store.dart';
 export 'src/store/db_cache_store.dart';
 export 'src/store/file_cache_store.dart';
+export 'src/store/mem_cache_store.dart';