+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.4.0
+
+- Add MemCacheStore. Cache with LRU strategy on RAM.
+- Re-add 'exists' store method.
+- Multiple fixes on maxStale (purge, get/set)
+- Various cleanups.
+
 ## 0.3.0
 
 - Add decrypt/encrypt methods to CacheOptions (no encryption by default. Private directories for stores are prefered for common usage).